+++ src/blocks/python/edit.tsx
@@ -6,7 +6,8 @@ import { PanelBody, PanelRow, ResizableBox, TextControl, ToggleControl, ToolbarB
 import { useEffect, useState } from '@wordpress/element'
 import { __ } from '@wordpress/i18n'
 import { edit } from '@wordpress/icons'
-import { Box, Button, ExtendedView, Group, SetupEditorEnv, SourceSelect, StateProvider, Text, useGlobalState } from '../../components'
+import { Box, Button, Group, SetupEditorEnv, SourceSelect, StateProvider, Text, useGlobalState } from '../../components'
+import { ExtendedView } from '../../components/ExtendedView'
 import { HidingWrapper } from '../../utils'
 import json from './block.json'
 import { Attributes } from './index'
