+++ src/components/index.tsx
@@ -5,10 +5,3 @@ export * from './CodeEditor'
 export * from './InseriThemeProvider'
 export * from './SetupEditorEnv'
 export * from './StateProvider'
-export * from './DetailViewBody'
-export * from './ContentTypeSelect'
-export * from './SelectWithAction'
-export * from './Cytoscape'
-export * from './ExtendedView'
-export * from './TopBar'
-export type * from './WorkerActions'
